+++ src/assets/RegisterAssetAssets/register-asset.png
