+++ src/components/SupplyChainComponents/SupplyChainComponents.js
@@ -11,10 +11,10 @@ import {
 const { height, width } = Dimensions.get('window');
 import Icon from "react-native-vector-icons/Entypo";
 const hercpngIcon = require('../assets/icons/hercIcon.png');
-import { HercTextFieldWithLabel } from "../components/SharedComponents";
+import { HercTextFieldWithLabel } from "../SharedComponents";
 // import Icon from 'react-native-vector-icons/FontAwesome';
-import ColorConstants from "../assets/ColorConstants";
-import { widthPercentageToDP, heightPercentageToDP } from '../assets/responsiveUI';
+import ColorConstants from "../../assets/ColorConstants";
+import { widthPercentageToDP, heightPercentageToDP } from '../../assets/responsiveUI';
 
 
 export function TransInfoCard(props) {
