+++ src/components/modals/confModalStyles.js
@@ -1,6 +1,6 @@
 import { StyleSheet, Platform } from "react-native";
-import ColorConstants from "../constants/ColorConstants";
-import { widthPercentageToDP, heightPercentageToDP } from "../assets/responsiveUI";
+import ColorConstants from "../../constants/ColorConstants";
+import { widthPercentageToDP, heightPercentageToDP } from "../../assets/responsiveUI";
 
 module.exports = StyleSheet.create({
