+++ src/screens/Testing/SupplyChainMetrics.js
@@ -15,7 +15,7 @@ import assets from "../../components/TesterAssets";
 import Header from "../../components/Headers/Header";
 
 // import Header from "../../components/Headers/Header";
-import { TransInfoCard } from "../../components/SupplyChainComponents";
+import { TransInfoCard } from "../../components/SupplyChainComponents/SupplyChainComponents";
 
 import {
     BigYellowButton, // props =  buttonName, onPress
